refactor(pages/home_page): :recycle: Moved FABs to their own files
